Don't copy the epigenome ids file if it already exists when installed source data.

+++ install_source_data.sh
@@ -22,7 +22,9 @@ function Roadmap_Epigenomics_15_state_model() {
     echo "Installing Roadmap_Epigenomics_15_state_model"
     mkdir -p $1/source_data/roadmap_r9/15_state_model/{raw,bed}
     mkdir -p $1/source_data/roadmap_r9/meta_data/
-    cp $CRAFT_GP_DATA/roadmap_r9/meta_data/roadmap_consolidated_epigenome_ids.csv $1/source_data/roadmap_r9/meta_data/
+    if [ ! -e $1/source_data/roadmap_r9/meta_data/roadmap_consolidated_epigenome_ids.csv ] ; then
+	cp $CRAFT_GP_DATA/roadmap_r9/meta_data/roadmap_consolidated_epigenome_ids.csv $1/source_data/roadmap_r9/meta_data/
+    fi
     # Get files
     cd $1/source_data/roadmap_r9/15_state_model/raw/
     wget http://egg2.wustl.edu/roadmap/data/byFileType/chromhmmSegmentations/ChmmModels/coreMarks/jointModel/final/all.dense.browserFiles.tgz
